Dirs.c: Cleanups and refactoring.
Split the core of mkdirs() out into a new function called mkdirs_chdir.
This allows an application to stay in the newly-created directory if
they want to access it without fear of symlink attacks.

Converted mkdirs() and mkdirs_chdir() to use const path inputs and
outputs (with an internal copy operation). This isn't strictly
necessary, but should help block against usage errors later down
the road.